Added int conversion for n_units

+++ overlapping_features.py
@@ -72,7 +72,7 @@ def max_fi_per_feature(pwr_per_feature, n_units_per_feature, dims,
 
 def estimate_mse(pwrs, n_units, dims, div=2, **kwargs):
     pwrs = np.array(pwrs)
-    n_units = np.array(n_units)
+    n_units = np.array(n_units).astype(int)
     out_full, c_full = rfm.emp_rf_decoding(pwrs, n_units, dims, **kwargs)
     out_div, c_div = rfm.emp_rf_decoding(pwrs/div, int(n_units/div),
                                          int(dims/div), **kwargs)
